Change to work in login shells
In a login shell, ```$0``` is e.g. ```-bash``` with a ```-``` at the beginning. This is a problem for the ME expansion, because ```ME=${ME:-`basename $0`}``` becomes ```ME=${ME:-`basename -bash`}```, which fails because basename takes -bash to be flags.  The solution is to use a double-dash ```--``` before the string, so basename doesn't interpret it as flags.